# Conflicts:
#	home-radar-backend/src/main/kotlin/api/PropertyController.kt
#	home-radar-backend/src/main/kotlin/domain/Property.kt
#	home-radar-backend/src/main/kotlin/service/PropertyService.kt
#	home-radar-backend/src/main/kotlin/web/extensions/ResponseUtils.kt
#	home-radar-frontend/src/app/core/services/property.service.ts
#	home-radar-frontend/src/app/pages/home/home.page.ts

+++ home-radar-backend/src/main/kotlin/web/extensions/ResponseUtils.kt
@@ -27,7 +27,7 @@ fun Property.toResponse() = PropertyResponse(
     bedrooms = bedrooms,
     imageUrl = imageUrl,
     neighborhood = neighborhood,
-    price = price.toDouble(),
+    price = price,
     yearBuilt = yearBuilt,
     ownerId = owner.id
 )
